requirements.txt
py-clob-client>=0.20.0
web3>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
matplotlib>=3.7.0
pandas>=2.0.0
aiohttp>=3.9.0
python-dateutil>=2.8.0
colorama>=0.4.6
tabulate>=0.9.0
urllib3>=2.0.0
